Moved to new date picker, add mockk/kotlintest/derby implementations. Just a few tests to fix and need to make the date pickers nicer... then we're done!

+++ Darts/src/main/kotlin/dartzee/utils/DevUtilities.kt
@@ -20,7 +20,7 @@ object DevUtilities
 
         val choice = DialogUtil.showInput("Delete Game", "Select Game ID", gameIds, gameIds[0]) ?: return
 
-        val gameId = choice as Long
+        val gameId = choice
         purgeGame(gameId)
     }
     private fun getAllGameIds(): Array<Long>?
